Fix the check of the post method and updating the okhttp jar file.
- In the FetchPlugin for Android, it is used postBody variable to check
if it is necessary to send a post (by checking if it is null). The
problem is: from JS, when null is passed, it becomes a valid
(not null) string containing the word "null". As a consequence, the
condition (postBody != null), in the line 55 of the current
FetchPlugin.java, will always be true. This commit fix this issue by
also checking if the postBody variable does not contain the word "null".

- The okhttp project released a new version of their okhttp library. In
this case, this commit updates the okhttp to the latest version and
makes the proper update in the plugin.xml.