Remove the reference to the AFHTTPRequestOperation.m/h in the plugin.xml
After @aporat's update to use AFNetwoking 3.0.1, the files
AFHTTPRequestOperation.m/h are not necessary, but they are still
referred in the plugin.xml. This commit removes these references because
it was breaking the build of this plugin.

In addition, I add a LOG message to print the HTTP response code because
it facilitates to know if it is the expected http code while coding,
specially considering when developing based on rest APIs.