Delete src/android/libs and automatic download libs from maven repo.
This plugin has two java libs dependencies: okio-1.6.0 and okhttp-2.7.0.
Everytime a new release is provided by the okhttp project, we had to
download the new .jar file, change the plugin.xml and remove the old jar
files.

This commit uses gradle to automatically downloads the dependencies from
a maven repository. Cordova supports gradle (in fact, it is the build
system used by Cordova). In this case, I create a file
named src/android/dependencies.gradle that specifies the dependencies
for the aforementioned libs and the proper version. In this way, once a
new version of the okhttp is provided by the okhttp project, we will
just need to update the dependencies.gradle to refer to the new version
and gradle will automatically download the jar files and includes in the
classpath of the project in the time of the plugin build.

Please, note that gradle just download the jar files of the okhttp and
okio just once, chaching them in a proper directory -- usually in the
~/.gradle, where other external dependencies for cordova (and any other
project that uses gradle) is also stored.